requirements.txt
Flask~=3.1.0
mysql-connector-python~=8.0.0

nltk~=3.9.1



PyPDF2~=3.0.1
python-docx~=1.1.2
